--- src/views/Map/Map.css
@@ -1,8 +1,0 @@
-#mapid {
-    position: absolute;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    top: 0;
-    z-index: -1;
-}
